fix/move utils to separate folder
